singular-agents/coborrower-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pdfplumber==0.10.3
python-dotenv==1.0.0
Werkzeug==3.0.1